feat(deepseek): share Files API uploader across chat completions & responses
