+++ app/src/main/java/com/ai/assistance/operit/api/chat/llmprovider/DeepseekProvider.kt
@@ -11,17 +11,9 @@ import com.ai.assistance.operit.data.model.ToolPrompt
 import com.ai.assistance.operit.util.ChatUtils
 import com.ai.assistance.operit.util.stream.Stream
 import okhttp3.OkHttpClient
-import okhttp3.MediaType.Companion.toMediaType
-import okhttp3.MultipartBody
-import okhttp3.Request
 import okhttp3.RequestBody
 import org.json.JSONArray
 import org.json.JSONObject
-import java.net.URL
-import java.util.Base64
-import java.util.concurrent.ConcurrentHashMap
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 
 /**
  * 针对DeepSeek模型的特定API Provider。
@@ -56,110 +48,17 @@ class DeepseekProvider(
         thinkingOptionId = thinkingOptionId
     ) {
     private val configuredApiEndpoint = apiEndpoint
-    // 基类中的 client / apiKeyProvider / customHeaders 为 private，这里保存自有引用供 Files API 上传使用
-    private val httpClient = client
-    private val keyProvider = apiKeyProvider
-    private val headerOverrides = customHeaders
+    /** DeepSeek Files API 图片上传器（与 Responses 协议线共用的抽取实现） */
+    private val fileUploader = DeepseekFileUploader(apiEndpoint, apiKeyProvider, client, customHeaders)
 
-    /** 图片 id -> DeepSeek Files API file_id 的会话级缓存（避免同一图片重复上传） */
-    private val uploadedImageFileIds = ConcurrentHashMap<String, String>()
-
-    /** DeepSeek Files API 支持上传的图片 MIME 类型 */
-    private val supportedUploadMimeTypes = setOf("image/jpeg", "image/png", "image/gif", "image/webp")
-
-    /** DeepSeek Files API 单文件上传上限（64 MiB） */
-    private val maxUploadBytes = 64L * 1024L * 1024L
-
-    /**
-     * 发送前把消息中的图片上传至 DeepSeek Files API 换取 file_id。
-     * 上传失败的图片保持原样，构建请求时回退为 base64 内嵌（见 buildImageContentPart）。
-     */
+    /** 发送前把消息中的图片上传至 DeepSeek Files API 换取 file_id，失败回退 base64 */
     override suspend fun prepareMediaForRequest(context: Context, chatHistory: List<PromptTurn>) {
-        if (!supportsVision) return
-        val seenIds = mutableSetOf<String>()
-        val imageLinks = mutableListOf<ImageLink>()
-        for (turn in chatHistory) {
-            MediaLinkParser.extractImageLinks(turn.content).forEach { link ->
-                if (seenIds.add(link.id)) imageLinks.add(link)
-            }
-        }
-        for (link in imageLinks) {
-            if (uploadedImageFileIds.containsKey(link.id)) continue
-            if (link.mimeType.lowercase() !in supportedUploadMimeTypes) continue
-            val bytes =
-                runCatching { Base64.getDecoder().decode(link.base64Data) }.getOrNull() ?: continue
-            if (bytes.isEmpty() || bytes.size > maxUploadBytes) continue
-            val fileId = uploadImageToFilesApi(bytes, link.mimeType, link.id)
-            if (fileId != null) {
-                uploadedImageFileIds[link.id] = fileId
-                AppLogger.d("DeepseekProvider", "图片 ${link.id} 已上传至 DeepSeek Files API: $fileId")
-            }
-        }
-    }
-
-    /** 上传单张图片到 DeepSeek Files API，返回 file_id；失败返回 null */
-    private suspend fun uploadImageToFilesApi(
-        bytes: ByteArray,
-        mimeType: String,
-        linkId: String
-    ): String? {
-        val filesEndpoint = buildFilesEndpoint() ?: return null
-        return runCatching {
-            val body =
-                MultipartBody.Builder()
-                    .setType(MultipartBody.FORM)
-                    .addFormDataPart("purpose", "user_data")
-                    .addFormDataPart(
-                        "file",
-                        "image_${linkId.take(24)}.${extensionForMime(mimeType)}",
-                        RequestBody.create(mimeType.toMediaType(), bytes)
-                    )
-                    .build()
-            val builder = Request.Builder().url(filesEndpoint).post(body)
-            applyAuthenticationHeaders(builder, keyProvider.getApiKey().trim())
-            headerOverrides.forEach { (key, value) ->
-                if (!key.equals("Content-Type", ignoreCase = true)) builder.addHeader(key, value)
-            }
-            withContext(Dispatchers.IO) { httpClient.newCall(builder.build()).execute() }.use { response ->
-                if (!response.isSuccessful) {
-                    AppLogger.w(
-                        "DeepseekProvider",
-                        "Files API 上传失败 HTTP ${response.code}: ${response.body?.string()?.take(200)}"
-                    )
-                    return@use null
-                }
-                val json = JSONObject(response.body?.string().orEmpty())
-                json.optString("id").takeIf { it.startsWith("file-") }
-            }
-        }.getOrElse { e ->
-            AppLogger.w("DeepseekProvider", "Files API 上传异常: ${e.message}")
-            null
-        }
-    }
-
-    /** 由聊天 API 端点推导 Files API 上传地址（官方为 https://api.deepseek.com/files） */
-    private fun buildFilesEndpoint(): String? {
-        val raw = configuredApiEndpoint.trim().removeSuffix("#").trim()
-        return runCatching {
-            val url = URL(raw)
-            val portPart = if (url.port in 1..65535) ":${url.port}" else ""
-            "${url.protocol}://${url.host}${portPart}/files"
-        }.getOrNull()
-    }
-
-    private fun extensionForMime(mimeType: String): String {
-        return when (mimeType.lowercase()) {
-            "image/jpeg" -> "jpg"
-            "image/png" -> "png"
-            "image/gif" -> "gif"
-            "image/webp" -> "webp"
-            else -> "bin"
-        }
+        fileUploader.prepareMedia(supportsVision, chatHistory, "DeepseekProvider")
     }
 
     /** 已有 file_id 时用 DeepSeek 文件块引用，否则回退基类的 base64 内嵌 */
     override fun buildImageContentPart(link: ImageLink): JSONObject {
-        val fileId = uploadedImageFileIds[link.id]
+        val fileId = fileUploader.fileIdFor(link.id)
         if (fileId != null) {
             return JSONObject().apply {
                 put("type", "file")
@@ -168,7 +67,6 @@ class DeepseekProvider(
         }
         return super.buildImageContentPart(link)
     }
-
     companion object {
         fun create(
             config: ModelConfigData,
@@ -718,6 +616,30 @@ private class DeepseekResponsesProvider(
 ) {
     override val useResponsesApi: Boolean = true
     override val bufferResponsesOutputTextUntilItemDone: Boolean = true
+    /** DeepSeek Files API 图片上传器（与 Chat Completions 协议线共用的抽取实现） */
+    private val fileUploader =
+        DeepseekFileUploader(responsesApiEndpoint, apiKeyProvider, client, customHeaders)
+
+    /** 发送前把消息中的图片上传至 DeepSeek Files API 换取 file_id，失败回退 base64 */
+    override suspend fun prepareMediaForRequest(context: Context, chatHistory: List<PromptTurn>) {
+        fileUploader.prepareMedia(supportsVision, chatHistory, "DeepseekResponsesProvider")
+    }
+
+    /**
+     * Responses 线同样先输出 DeepSeek 文件块，随后由
+     * OpenAIResponsesPayloadAdapter.convertMessageContentForResponses 转换为
+     * Responses 协议的 input_image + file_id 引用；未上传成功则回退 base64。
+     */
+    override fun buildImageContentPart(link: ImageLink): JSONObject {
+        val fileId = fileUploader.fileIdFor(link.id)
+        if (fileId != null) {
+            return JSONObject().apply {
+                put("type", "file")
+                put("file_id", fileId)
+            }
+        }
+        return super.buildImageContentPart(link)
+    }
 
     override fun isResponsesCommentaryMessage(item: JSONObject): Boolean {
         return item.optString("phase", "").trim().equals("commentary", ignoreCase = true)
